DTO projection

+++ src/main/java/dayp308/chatroom/entity/view/comment/CommentView.java
@@ -17,7 +17,7 @@ public class CommentView {
     protected String commentBody;
     protected Instant createTime;
     protected Long parentCommentId;
-    protected Long likes;
+    protected Integer likes;
 
     @Override
     public boolean equals(Object o) {
